Implement another chunk of opcode actions
Down to 5 left, including a few of the more involved one.  I also
settled on a representation for the key-pressed data, as just a
simple vector of booleans.  The event loop can update this each
program tick and "blocking" was more simply represented by looping
over the same instruction repeatedly rather than using some mechanism
external to the emulator.