Finish all but 1 of the opcodes, add docs and tidy
Other than the inner workings of drawSprite, all of the opcodes are
written and in place.  Rather than writing a number of tests myself,
I am waiting to run some test CHIP-8 programs to help me find logic
errors instead for less of a time investment.

The body of the main loop has been written out in some psuedocode,
that will require pulling in SDL to handle keyboard events and
actually rendering the screen data.

I cleaned up a lot of repetition in the decoding function, and added
missing docstrings for ProgramState and it's fields to make it easier
to remember specification details later.