Clean up .cabal file more
Remove CHIP8 prefixes to all modules, they are already grouped by folders
and I do not expect any shared code to live in chip8* build targets.  This
also simplifies the cabal file, because where to put which piece and which
needed a dot or directory separator, which parts needed an extra "main-is"
flag was kind of a pain.