Minor rewrites/comments to wrap up disassembler
It is of no use right now on full programs, but could still be useful for
analyzing chunks of known program instruction data.  With programs being able
to store data and jump things might not align exactly as expected.  Maybe an
upgrade later on when the main emulator work is in place.